Some cleanup

+++ src/eval.cpp
@@ -13,7 +13,7 @@ Eval::Eval() {
                 return a * b;
             };
         eval_map_["/"] = [](int64_t a, int64_t b)->int64_t {
-                return a / b;
+                return (int64_t) a / b;
             };
 }
